fix: ledger usb connection error
